Rename setdatasource.py to setDataSource.py
